Fix transition stuck in AOD and keyguard clock visible after unlock
It happens when rapidly switching between AOD and Hub, and transition
interruption not handled as expected. During Blank(AOD)->Communal(Hub)
scene transition, if we press the power button to interrupt it, the
CommunalSceneInteractor requests a Communal->Blank transition but STL
will start a reversed Blank->Communal transition (currentScene=Blank).
It causes keyguard state not updated correctly and finally stuck in AOD.

Fix is to update ktf state when a reversed transition is detected due
to interruption in handleTransition and handleIdle.

Fixes: b/395041829
Test: rapidly switch between AOD<->Hub by power button then unlock
Test: atest CommunalSceneTransitionInteractorTest
Flag: EXEMPT bugfix
Change-Id: Ib177e33a77db1157613db5e89cf279b26c4ee468