[1/N] Turn OomAdjuster into an abstract base class
And create OomAdjusterLegacyImpl, which overrides the same methods that
OomAdjusterModernImpl overrides (enabling a clear set of methods in
OomAdjuster to make abstract).

The main methods the OomAdjuster implementations override are the
`performUpdateOomAdj.*` methods. These methods are the entry points into
the differing update algorithm of each impl.

This CL moves the bare minimum amount of code to the new
OomAdjusterLegacyImpl to maintain the current behaviors. The next
several CLs with migrate more functionality out of the base class until
all implementation specific behavior lives in the appropriate file.

Flag: EXEMPT pure refactor
Test: atest MockingOomAdjusterTests
Test: builds and flashes
Bug: 306731120
Change-Id: I70f4b3dacae1363171fa95b421a90d502c60e5d0