Remove more redundant fillInStackTrace calls
Throwable's constructor already calls it, and it is quite expensive.
All removed calls were redundant.

We often see suspension timeouts triggered by stack trace generation.
Much of the cost is in paging in rarely used information, so clearly
redundant calls like these are likely to be less expensive. But they
still seem worth removing.

This continues an exercise started with ag/30132636 .

Note that a few of these don't appear to be on exceptional paths,
and it may be worth removing the exception construction as well.
I leave that to people more familiar with this code.

Bug: 407383437
Test: Treehugger
Flag: EXEMPT bugfix
Change-Id: I46630d4ac860141f1142af509d24ee33568068d8