Add nativeService attribute to the service tag
This adds a new attribute "nativeService" and assocated properties to
the service tag in AndroidManifest.xml.

The attribute nativeService is used to indicate that the service is
intended to run on a native process.

The properties PROPERTY_NATIVE_SERVICE_LIB_NAME and
PROPERTY_NATIVE_SERVICE_FUNC_NAME are used to configure the library name
and entrypoint function name respectively.

Bug: 431902161
Test: Build and install a test app
Flag: android.os.native_framework_prototype
Change-Id: I65b69cf939a828cb002edec904cb84391be0a241