Run SelectionToolbarRenderService in SystemUi
This adds an implementation of the selection toolbar render service into
the system ui package. Most of the implementation remains in
framework.jar for easy maintenance between the two processes we may run
it in. When proven successful the full implementation is planned to be
moved into system ui and the rest converted to kotlin.

Bug: 363318732
Test: TextViewIntegrationTest TextViewMouseInteractionTest and manual
Flag: android.permission.flags.use_system_selection_toolbar_in_sysui
Change-Id: I762ebff532f4f85a7dd30670ca8cbf6be3c41b40