Make animation-background rel-z instead of reparented
For split, there were some situation where the animation-background
would be placed above (in Z) one of the split sides during
animation. This was originally resolved by putting the layer into
the root TDA. While this works, it relies on the finish transaction
being applied to clean it up. There are some situations where this
may not happen (for example if sysui process crashes) which can
leave the surface around "forever".

The transition root, however, is managed by Core so will be
cleaned up if sysui crashes, so it's more reliable if we can
leave the background as a child of that.

While neither of these is a very scalable solution, for the short
term: since the original issue was about z-order, we can replace
reparenting with rel-z so it will get cleaned-up with the
transition root.

Bug: 384435414
Test: launch task into other split to ensure nothing visual regressed
Flag: EXEMPT bugfix
Change-Id: Ied07853b17025829191dadb061960474b5d57890