Cache isLauncherUnderneath and update when finished in not-GONE.
This requires a binder call to ATMS#getTasks, which was happening 1-2 times during unlock and adding several ms to startTransition.

Since it's very unlikely that the app underneath will change while we're locked (something would have to crash, or be launched via alarm/adb), we can simply update this whenever we finish a transition away from GONE. Worst case, if we're wrong, then either launcher will just play the normal app-window animation, or launcher will play the in-window animation when it's not visible, neither of which are a big deal.

Bug: 278086361
Flag: com.android.systemui.keyguard_wm_state_refactor
Test: perfetto trace, see only one call to getTasks() and it's not during unlock CUJ
Test: atest InWindowLauncherUnlockAnimationInteractorTest
Change-Id: I36684fc32d8396f3bd9d0c088f544b4209317595