Remove the AnrTimer freezer feature
The AnrTimer can freeze ANR'ed processes until system server is ready
to take a stack dump.  The process is unfrozen when the ANR report is
generated.  The process is also unfrozen in a finalizer, in case
system_server fails to take an ANR report.  However, the unfreeze
operation can induce a finalizer timeout.

This change removes all code associated with freezing ANR'ed
processes.  The flag is removed as well.

Flag: EXEMPT flag-deletion
Bug: 406028115
Test: atest
 * FrameworksServicesTests:com.android.server.am
 * CtsAppTestCases
 * FrameworksServicesTests:AnrTimerTest
Change-Id: Ia054816604daac20a9edb216ff0e0c1a104c5af7