Add numpad support for mouse keys
Add support for mouse keys when used with keyboards with a numpad.
The default key bindings for mouse keys is still the primary keys
(since not all keyboards come with numpads).
If a user chooses to use numpads for mouse keys, we check if all the
numpad key bindings exist on the keyboard, If they exist, the user can
use mouse keys with the numpad, if not, it defaults back to the
primary keys.
If the user has multiple keyboards connected, and has selected to use
numpad for mouse keys, they will only be able to use the numpad for the
keyboards which support numpad, and primary keys with the keyboards that
don't.
Also update access to the mouse keys enums in a more efficient way.

Bug: 369779909
Test: atest FrameworksServicesTests:MouseKeysInterceptorTest
Flag: com.android.server.accessibility.enable_mouse_key_enhancement
Change-Id: Ia6d68730ead63a49ab7d5deb32f6b34e799af2c6