Init underlay from content capture services.
As the first step in system UI(ag/32801123), we are having a broadcast listener to control the lifecycle of underlay. The content capture service is responsible for detecting the lifecycle of the target apps and create/destroy the underlay accordingly. Will migrate to a binder aidl later.

Bug: 403422950
Flag: android.view.contentcapture.flags.enable_system_ui_underlay
Change-Id: I7c0020952a50ea89121c301f4942d68ac2244ea5
Test: atest CtsContentCaptureServiceTestCase