Fix ScreenUnblocker message abnormal waiting
On the Xiaomi Flip device, we found that when the device is unfolded or
folded with the screen on, the display information may be updated
immediately because DisplayContent.getLastHasContent returns false.
Then, DeferredDisplayUpdater.waitForTransition is called. At this time,
we also try to wait for the execution of the screen - switching
Transition. However, in fact, it is not created at all, so there is an
abnormal waiting situation.

For this kind of corner case, we perform a check after the display
information is updated. If the Transition is not created at this time,
we should not let it continue to wait for the Transition.

The reason why we don't use TransitionController.inTransition here is
that there may be a Transition for orientation change at this time,
which makes this judgment inaccurate. What we clearly need here is the
screen - switching Transition.

Bug: 406965640
Flag: EXEMPT bugfix
Test: Manual test on XiaoMi flip device

Change-Id: I0395f80ad645bfd3a63a9f80e6b8be869152f86a
Signed-off-by: Linnan Li <lilinnan@xiaomi.corp-partner.google.com>