Apply DisplayChange to display disconnect transition.
Modifies the display disconnect transition to include a DisplayChange
that flags the reparentDisplayId, the display that any reparented tasks
will move to.

This DisplayChange will be used by TransitionHandlers to determine any
additional changes to the transition that need to be made. Due to this,
reparent behavior in TaskDisplayArea has largely been reverted to what it was prior to I6ec4b9d305a6bd9497450328dad36bc53179ee7c (with the exception of some freeform organizer tasks not being removed).

Bug: 391652399
Test: Manual; confirm freeform tasks provide display change to Shell on disconnect
Flag: com.android.window.flags.enable_display_disconnect_interaction

Change-Id: I0e5a4ed3276a676b28a9470ab0c2739a7ef7e29e