Refactor widget event data to use dedicated class
This change adds AppWidgetEvent immutable data class which is used to
track the event data in the host and in the system service. The class
includes a Builder to create and update these objects. This simplifies
the implementation in both of these components and centralizes the
logic around merging multiple events.

Bug: 364655296
Test: atest CtsAppWidgetTestCases:WidgetEventsTest \
    FrameworksCoreTests:AppWidgetEventsTest
Flag: android.appwidget.flags.engagement_metrics

Change-Id: I8c4d01246ff989b2308e7ec52ec34bc0cc0f69a5