DragToDesktopTransitionHandler: Set DoNotPip on finishCallback.
When we are "cancelling" a drag-to-desktop by going to split instead,
because home is now on top, the dragged task is now sent to the back. If
the task can PIP, then Core will trigger respective lifecycle/visibility
checks and ends up putting it into PIP instead of enter-split state.

This sets a new WCT, and call #doNotPip on it, and send it as part of
the transition's finishCallback so the system has signal to not enter
the task into PiP in this case.

Bug: 406539545
Test: Manual. Have youtube, play video, drag to enter-split - it no
longer pips.
Flag: EXEMPT bugfix

Change-Id: I3a85f6eac62347c588d7f440edcfb04d5c699108