Check visible transient launch explicitly
When launching a visible transient task to front (a translucent
activity is on top), the transition participants contain the
transient launch task (reorder) but don't contain the activity
(no visibility change). Then hasVisibleTransientLaunch may be
missed to set. That causes the dependent logic of the flag not
to be called, e.g. update ime target.

So check from mTransientLaunches directly instead of depending on
transition participants.

Fix: 409453244
Flag: EXEMPT bugfix
Test: atest TransitionTests#testTransientLaunchWithTranslucentTask
Change-Id: I020228c6dfd58c43c94f586ad811ab183a98f041