Respect the hardware accelerated of ApplicationInfo
Enabling the flag enable_window_context_override_type led to
PlatformScenarioTests failure. The root cause is that setting
the parent window leads to the window itself and its sub-windows
respect if hardware accelerated is enabled, which disabled hardware
accelerated on the notification shade window,
and so did the ThreadedRenderer. That's why the view validaton in jank
monitor failed.

This CL set the hardware accelleration from ApplicationInfo, which
follows the mechanism in WindowManagerGlobal#addView.

Test: atest WindowContext ViewCaptureAwareWindowManagerTest
Fixes: 410504049
Flag: com.android.window.flags.enable_window_context_override_type
Change-Id: I9e2def6249729d490e0123d244db5ce061b16dc4