Add a new Recents transition state to signal a stop request.
When a recents animation is on its way to finishing, we want to get a
slightly earlier signal to certain parts (e.g. windowdecor) that it is
going to stop running soon. Instead of only sending a signal when the
transition clean up is happening, we will send another state to notify
that the transition is on its way to be done (e.g. finishInner is
already starting).

Bug: 410838192
Test: atest
Flag: com.android.window.flags.enable_input_layer_transition_fix
Change-Id: Ie183f96cac245c4f859bb74ed51b5d159de51547