Fix NotificationSettingsControllerTest
Test is dependent on ActivityManager.getCurrentUser. However, there is a mock object that returns userId which is not inline with getCurrentUser.
ActivitiyManager.getCurrentUser has been started to return a value different from 0, which is a default value for int and so mock userId. This CL fixes this issue by improving the mock

Fix: 411119789
Test: NotificationSettingsControllerTest
Flag: EXEMPT trivial test set up change
Change-Id: I291da83fdc14e20fae65f257b44e06eb000b8d43