Don't show dismissible keyguard in app pinning mode
On foldables, if the "Continue using apps on fold" is set
to "Swipe up to continue", when closing the device,
a dismissible keyguard will show.

If this setting is set, and there's a pinned app, instead
of showing the dismissible keyguard on fold, we keep the
screen on with the app displayed.
This keeps the app pinned and in the same security state as
when it was unfolded.

Bug: 404252173
Flag: EXEMPT bugfix
Test: manually (given settings described above), pin app, fold;
observe app still shows with screen on

Change-Id: I61e2014b9d21862859090707e5531a61bf966fc3