Fix expansion and heights for bundled notifications (and groups)
With this change expanding and collapsing groups and notifications
within bundles is possible and looks correct. The initial collapsed
state of bundled groups is a bit off and will be fixed in subsequent CL.

Bug: b/389839492
Test: added tests and manual verificion with notifications and groups
in the debug bundle. Tested all variations of expand and collapse.
Flag: com.android.systemui.notification_bundle_ui

Change-Id: I3ae911ce176fabae0135bd19674a455cc62fc94b