Add lock screen content description in coroutine
Adding lock screen content description in ag/32709985 is causing
performance regression seen in b/408731118 and b/408634675. Setting
the view's contentDescription in `setSnapshotBinding` is delaying the
UI frame. Reverted ag/32858690 and this change re-adds the contentDescription inside of coroutine to avoid blocking `setSnapshotBinding`.

This is a temporary solution to resolve regressions until we find a
better solution.

Bug: 408050048
Bug: 413403813
Flag: EXEMPT bug fix
Test: manual - verify Talkback lock screen label in lock screen
and no label with quick setting shade open

Change-Id: I1358e5f00ecb748baa91b78b253337e10faf9b0e