Setting a minimum dismiss velocity.
When dismissing a notification after magnetically detaching, the dismiss
animation can be too slow if the touch input lifts at zero velocity. To
remediate this issue, we se a minimum dismiss velocity when dismissing.
We also make sure that the dismissal occurs in the same direction that a
notification has been detached.

Test: Added unit tests
Test: manual. Verified that notifications dismiss at the minimum set
  velocity when releasing them at zero finger velocity after detaching.
Flag: com.android.systemui.magnetic_notification_swipes
Bug: 414432406
Change-Id: I428bc9b047ba4ceb1d9cbd79824e9e13ff444c9f