Fix HUN disappear flickering by canceling animators after clipping
Root cause: With physicalNotificationMovement enabled, the HUN y-position
spring animation lasts longer than its content/clipping animation. This
causes the clipping rectangle to be reset prematurely while the NPV is still
visible due to the ongoing y-translation, resulting in a visual glitch.

Fix: Advance any running animators to the end for the HUN disappearing,
after its content animation has finished, because at this point the view is
fully clipped off, and it shouldn't be visible anyways.
This results in calling NSSL#onChildAnimationFinished(), which triggers
the natural way of finalizing the HUN disappear animation, which is
setting headsUpAnimationAway to false, which hides the Notification Panel.

Fixes: 406254389
Test: observe the HUN appear/disappear animations
Flag: com.android.systemui.physical_notification_movement
Change-Id: I6cdac864fd58db13ca8996f7d4425a740f366351